+++ src/Microsoft.AspNet.Mvc.Core/Filters/AuthorizeFilter.cs
@@ -14,12 +14,20 @@ namespace Microsoft.AspNet.Mvc
     /// </summary>
     public class AuthorizeFilter : IAsyncAuthorizationFilter, IOrderedFilter
     {
+        /// <inheritdoc />
         public int Order { get; set; }
 
+        /// <summary>
+        /// Policy name used for authorization
+        /// </summary>
         public string Policy { get; set; }
 
+        /// <summary>
+        /// User must be a member of one of these Roles if any are specified
+        /// </summary>
         public IEnumerable<string> Roles { get; set; }
 
+        /// <inheritdoc />
         public virtual async Task OnAuthorizationAsync([NotNull] AuthorizationContext context)
         {
             var httpContext = context.HttpContext;
@@ -54,11 +62,17 @@ namespace Microsoft.AspNet.Mvc
             }
         }
 
+        /// <summary>
+        /// Returns true if there is an IAllowAnonymous filter
+        /// </summary>
+        /// <param name="context"></param>
+        /// <returns></returns>
         protected virtual bool HasAllowAnonymous([NotNull] AuthorizationContext context)
         {
             return context.Filters.Any(item => item is IAllowAnonymous);
         }
 
+        /// <inheritdoc />
         protected virtual void Fail([NotNull] AuthorizationContext context)
         {
             context.Result = new HttpStatusCodeResult(401);
