--- test/Microsoft.AspNet.Mvc.Core.Test/Filters/AuthorizeFilterTestsBase.cs
@@ -1,65 +1,0 @@
-// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using System.Linq;
-using System.Security.Claims;
-using Microsoft.AspNet.Http;
-using Microsoft.AspNet.Routing;
-using Microsoft.Framework.DependencyInjection;
-using Microsoft.Framework.DependencyInjection.Fallback;
-using Moq;
-
-namespace Microsoft.AspNet.Mvc.Core.Test
-{
-    public class AuthorizeFilterTestsBase
-    {
-        protected AuthorizationContext GetAuthorizationContext(Action<ServiceCollection> registerServices, bool anonymous = false)
-        {
-            var validUser = new ClaimsPrincipal(
-                new ClaimsIdentity(
-                    new Claim[] {
-                        new Claim("Permission", "CanViewPage"),
-                        new Claim(ClaimTypes.Role, "Administrator"),
-                        new Claim(ClaimTypes.Role, "User"),
-                        new Claim(ClaimTypes.NameIdentifier, "John")},
-                        "Basic"));
-
-            validUser.AddIdentity(
-                new ClaimsIdentity(
-                    new Claim[] {
-                        new Claim("Permission", "CupBearer"),
-                        new Claim(ClaimTypes.Role, "Token"),
-                        new Claim(ClaimTypes.NameIdentifier, "John Bear")},
-                        "Bearer"));
-
-            // ServiceProvider
-            var serviceCollection = new ServiceCollection();
-            if (registerServices != null)
-            {
-                registerServices(serviceCollection);
-            }
-
-            var serviceProvider = serviceCollection.BuildServiceProvider();
-
-            // HttpContext
-            var httpContext = new Mock<HttpContext>();
-            httpContext.SetupGet(c => c.User).Returns(anonymous ? null : validUser);
-            httpContext.SetupGet(c => c.RequestServices).Returns(serviceProvider);
-
-            // AuthorizationContext
-            var actionContext = new ActionContext(
-                httpContext: httpContext.Object,
-                routeData: new RouteData(),
-                actionDescriptor: null
-                );
-
-            var authorizationContext = new AuthorizationContext(
-                actionContext,
-                Enumerable.Empty<IFilter>().ToList()
-            );
-
-            return authorizationContext;
-        }
-    }
-}
